+++ Oleander.StrResGen/src/Logger.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using Microsoft.Extensions.Logging;
+
+namespace Oleander.StrResGen;
+
+internal class Logger : ILogger
+{
+
+
+
+    public void Log<TState>(LogLevel logLevel, EventId eventId, TState state, Exception? exception, Func<TState, Exception?, string> formatter)
+    {
+       
+    }
+
+    public bool IsEnabled(LogLevel logLevel)
+    {
+        return true;
+    }
+
+    public IDisposable? BeginScope<TState>(TState state) where TState : notnull
+    {
+        return NullScope.Instance;
+    }
+}
