+++ Oleander.StrResGen/src/NullScope.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+namespace Oleander.StrResGen;
+
+internal sealed class NullScope : IDisposable
+{
+    public static NullScope Instance { get; } = new NullScope();
+
+    private NullScope()
+    {
+    }
+
+    /// <inheritdoc />
+    public void Dispose()
+    {
+    }
+}
